total reshuffle into cohortStages
