+++ cpg_workflows/targets.py
@@ -129,6 +129,32 @@ class Cohort(Target):
         """Unique target ID"""
         return self.name
 
+
+    def write_ped_file(
+        self, out_path: Path | None = None, use_participant_id: bool = False
+    ) -> Path:
+        """
+        Create a PED file for all samples in the whole cohort
+        PED is written with no header line to be strict specification compliant
+        """
+        datas = []
+        for sample in self.get_samples():
+            datas.append(
+                sample.pedigree.get_ped_dict(use_participant_id=use_participant_id)
+            )
+        if not datas:
+            raise ValueError(f'No pedigree data found for {self.name}')
+        df = pd.DataFrame(datas)
+
+        if out_path is None:
+            out_path = self.analysis_dataset.tmp_prefix() / 'ped' / f'{self.alignment_inputs_hash()}.ped'
+
+        if not get_config()['workflow'].get('dry_run', False):
+            with out_path.open('w') as fp:
+                df.to_csv(fp, sep='\t', index=False, header=False)
+        return out_path
+
+
     def get_datasets(self, only_active: bool = True) -> list['Dataset']:
         """
         Gets list of all datasets.
