+++ main.py
@@ -17,8 +17,8 @@ from cpg_workflows.stages.cram_qc import CramMultiQC
 from cpg_workflows.stages.gvcf_qc import GvcfMultiQC
 from cpg_workflows.stages.fastqc import FastQCMultiQC
 from cpg_workflows.stages.seqr_loader import MtToEs, AnnotateDataset, DatasetVCF
-from cpg_workflows.stages.gatk_sv import AnnotateVcf
-from cpg_workflows.stages.gatk_sv_single_sample import CreateSampleBatches
+from cpg_workflows.stages.gatk_sv.gatk_sv import AnnotateVcf
+from cpg_workflows.stages.gatk_sv.gatk_sv_single_sample import CreateSampleBatches
 from cpg_workflows.stages.stripy import Stripy
